+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Please view this file on the master branch, on other branches it's out of date.
 
 v1.6.1 (unreleased)
+  - Improved detection of the Google+ username
   - Media Gallery: Ajax driven rebuild thumbnails / resize display images
   - Media Gallery: removed config.php from distribution - rename .dist on install
   - Media Gallery: Use HTML5 playback for mp3 files